refactor(parametrage): renommage parametres_generaux en configuration_globale
- Renommage fichiers: utils_parametres_generaux.py -> utils_configuration_globale.py
- Renommage fichiers: parametrage/forms/parametres_generaux.py -> configuration_globale.py
- Renommage fichiers: parametrage/views/parametres_generaux.py -> configuration_globale.py
- Mise à jour imports dans core/views/base.py, core/views/menu.py, core/utils/utils_permissions.py
- Mise à jour URLs dans parametrage/urls.py
- Renommage variables locales et paramètres de fonctions
