+++ noethysweb/core/views/menu.py
@@ -8,8 +8,8 @@ from django.urls import reverse_lazy
 from django.conf import settings
 
 
-def GetMenuPrincipal(parametres_generaux=None, organisateur=None, user=None, force_permissions=False):
-    parametres_generaux = parametres_generaux or {}  # Fallback to an empty dictionary
+def GetMenuPrincipal(configuration_globale=None, organisateur=None, user=None, force_permissions=False):
+    configuration_globale = configuration_globale or {}  # Fallback to an empty dictionary
 
     menu = Menu(titre="Menu principal",  user=user, force_permissions=force_permissions)
